Refactor ins
