+++ hive/lib/hive.dart
@@ -27,7 +27,7 @@ export 'src/object/hive_storage_backend_preference.dart';
 export 'src/registry/type_adapter.dart';
 export 'src/registry/type_registry.dart';
 
-export 'src/isolate/isolate_name_server/isolate_name_server.dart';
+export 'src/isolate/isolate_name_server.dart' if (dart.library.js_interop) '';
 export 'src/isolate/isolated_hive.dart';
 export 'src/isolate/isolated_box.dart';
