--- hive/lib/src/isolate/isolate_name_server/impl/isolate_name_server_stub.dart
@@ -1,14 +1,0 @@
-/// Stub
-abstract class IsolateNameServer {
-  /// Constructor
-  const IsolateNameServer();
-
-  /// Stub
-  dynamic lookupPortByName(String name);
-
-  /// Stub
-  bool registerPortWithName(dynamic port, String name);
-
-  /// Stub
-  bool removePortNameMapping(String name);
-}
