--- hive/lib/src/isolate/isolate_name_server/isolate_name_server.dart
@@ -1,2 +1,0 @@
-export 'impl/isolate_name_server_vm.dart'
-    if (dart.library.js_interop) 'impl/isolate_name_server_stub.dart';
