+++ hive/test/tests/isolate/isolated_hive_test.dart
@@ -6,7 +6,7 @@ import 'package:hive_ce/hive.dart';
 import 'package:hive_ce/src/isolate/isolated_hive_impl/isolated_hive_impl.dart';
 import 'package:test/test.dart';
 
-import '../../integration/isolate_test.dart';
+import '../../util/ins/test_ins.dart';
 import '../../util/is_browser/is_browser.dart';
 import '../common.dart';
