+++ hive/test/util/ins/test_ins.dart
@@ -1,0 +1,1 @@
+export 'test_ins_impl.dart' if (dart.library.js_interop) '';
